Add headset icon toggle (CMParts)
What:
Adds a setting that enables/disables the headset icon from displaying on the
status bar when a wired headset is plugged in.

Why:
Many users have requested this feature because the icon wastes space on their
status bar and is somewhat unnecessary. Also, other UI's for android do not
include a headset icon (Sense, for example), so users may be used to not
seeing the icon and prefer it that way. Users with this preference can now
disable the icon from ever appearing.

Change-Id: I1ef20a4cc8962a3ff5088e215a67dee1da53a02f